IE2YFRG_KOOT38UG/requirements.txt
# Packages in this file will be installed into the app environment
streamlit-nightly